protect all routes

+++ backend/controllers/goalsController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require('express-async-handler');
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 // @desc get goals
 // @route GET /api/goals
 // @access privete
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json(goals);
 });
 
@@ -20,6 +21,7 @@ const createGoal = asyncHandler(async (req, res) => {
   }
 
   const goal = await Goal.create({
+    user: req.user.id,
     text: req.body.text,
   });
 
@@ -36,6 +38,18 @@ const updateGoal = asyncHandler(async (req, res) => {
     throw new Error('goal not found');
   }
 
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error('user not found');
+  }
+
+  // make sure the loggedin user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('user not authorized');
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -50,6 +64,18 @@ const deleteGoal = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error('goal not found');
+  }
+
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error('user not found');
+  }
+
+  // make sure the loggedin user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('user not authorized');
   }
 
   await goal.deleteOne();
